outsources sending email to a separate file
